# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Smart City Digital Citizen Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transport/routes:
    get:
      operationId: listTransportRoutes
      tags: [transport]
      summary: List public transport routes
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [bus, metro, tram]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Transport routes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransportRoute"
  /transport/routes/{id}:
    get:
      operationId: getTransportRoute
      tags: [transport]
      summary: Get transport route details with stops
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transport route details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransportRouteDetail"
...
  /parks:
    get:
      operationId: listParks
      tags: [parks]
      summary: List all parks
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Parks list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Park"
  /parks/{id}:
    get:
      operationId: getPark
      tags: [parks]
      summary: Get park details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Park details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Park"

  # ─────────────────────────────── LIBRARIES ───────────────────────────────
...
  /libraries/books:
    get:
      operationId: listBooks
      tags: [libraries]
      summary: Search books across all libraries
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: libraryId
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Books list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookListResponse"
  /libraries/borrowings:
    get:
      operationId: listBorrowings
      tags: [libraries]
      summary: List user borrowings
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, returned, overdue]
      responses:
        "200":
          description: User borrowings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Borrowing"
    post:
      operationId: borrowBook
      tags: [libraries]
      summary: Borrow a book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BorrowInput"
      responses:
        "201":
          description: Book borrowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Borrowing"
  /libraries/borrowings/{id}/return:
    post:
      operationId: returnBook
      tags: [libraries]
      summary: Return a borrowed book
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Book returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Borrowing"

  # ─────────────────────────────── PAYMENTS ───────────────────────────────
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List payment history
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [property_tax, water_tax, parking, other]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Payment history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentListResponse"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get payment details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # ─────────────────────────────── NOTIFICATIONS ───────────────────────────────
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List user notifications
      parameters:
        - name: unread
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /feedback:
    get:
      operationId: listFeedback
      tags: [feedback]
      summary: List feedback (admin)
      parameters:
        - name: rating
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Feedback list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackListResponse"
    post:
      operationId: submitFeedback
      tags: [feedback]
      summary: Submit feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeedbackInput"
      responses:
        "201":
          description: Feedback submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"
...
  /admin/citizens:
    get:
      operationId: listCitizens
      tags: [admin]
      summary: List all citizens (admin)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Citizens list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CitizenListResponse"
  /admin/audit-logs:
    get:
      operationId: listAuditLogs
      tags: [admin]
      summary: List audit logs (admin)
      parameters:
        - name: action
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogListResponse"
  /admin/reports/revenue:
    get:
      operationId: getRevenueReport
      tags: [admin]
      summary: Get revenue report
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [7d, 30d, 90d, 1y]
            default: 30d
      responses:
        "200":
          description: Revenue report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueReport"
...
components:
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [total, page, limit, totalPages]
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [citizen, admin, super_admin]
        createdAt:
          type: string
          format: date-time
      required: [id, clerkId, email, firstName, lastName, role, createdAt]
...
    TransportRoute:
      type: object
      properties:
        id:
          type: string
        routeNumber:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [bus, metro, tram]
        origin:
          type: string
        destination:
          type: string
        frequency:
          type: integer
        operatingHours:
          type: string
        fare:
          type: number
        totalStops:
          type: integer
        isActive:
          type: boolean
      required: [id, routeNumber, name, type, origin, destination, frequency, operatingHours, fare, totalStops, isActive]
    TransportStop:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        sequence:
          type: integer
        latitude:
          type: number
        longitude:
          type: number
        arrivalTime:
          type: string
      required: [id, name, sequence, latitude, longitude, arrivalTime]
    TransportRouteDetail:
      type: object
      properties:
        route:
          $ref: "#/components/schemas/TransportRoute"
        stops:
          type: array
          items:
            $ref: "#/components/schemas/TransportStop"
      required: [route, stops]
...
    Park:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        area:
          type: number
        openingTime:
          type: string
        closingTime:
          type: string
        amenities:
          type: array
          items:
            type: string
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
      required: [id, name, description, address, latitude, longitude, area, openingTime, closingTime, amenities, isActive]

    # ─── Libraries ───
...
    Book:
      type: object
      properties:
        id:
          type: string
        libraryId:
          type: string
        title:
          type: string
        author:
          type: string
        isbn:
          type: ["string", "null"]
        genre:
          type: string
        publishedYear:
          type: ["integer", "null"]
        totalCopies:
          type: integer
        availableCopies:
          type: integer
        isAvailable:
          type: boolean
      required: [id, libraryId, title, author, genre, totalCopies, availableCopies, isAvailable]
    BookListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Book"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, pagination]
    Borrowing:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        bookId:
          type: string
        book:
          $ref: "#/components/schemas/Book"
        borrowedAt:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date
        returnedAt:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          enum: [active, returned, overdue]
        fineDue:
          type: number
      required: [id, userId, bookId, book, borrowedAt, dueDate, status, fineDue]
    BorrowInput:
      type: object
      properties:
        bookId:
          type: string
      required: [bookId]

    # ─── Payments ───
...
    Payment:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
          enum: [property_tax, water_tax, parking, other]
        referenceId:
          type: string
        amount:
          type: number
        paymentMethod:
          type: string
        receiptNumber:
          type: string
        status:
          type: string
          enum: [success, failed, pending, refunded]
        paidAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, userId, type, referenceId, amount, paymentMethod, receiptNumber, status, createdAt]
    PaymentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, pagination]

    # ─── Notifications ───
    Notification:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [complaint_update, tax_due, certificate_approved, garbage_pickup, parking_reminder, transport_alert, general]
        isRead:
          type: boolean
        referenceId:
          type: ["string", "null"]
        referenceType:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, userId, title, message, type, isRead, createdAt]
    NotificationListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, pagination]

    # ─── Feedback ───
    Feedback:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        serviceType:
          type: string
          enum: [complaint, tax, certificate, garbage, parking, transport, parks, libraries, general]
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: ["string", "null"]
        sentiment:
          type: ["string", "null"]
          enum: [positive, neutral, negative]
        createdAt:
          type: string
          format: date-time
      required: [id, userId, serviceType, rating, createdAt]
    FeedbackInput:
      type: object
      properties:
        serviceType:
          type: string
          enum: [complaint, tax, certificate, garbage, parking, transport, parks, libraries, general]
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
      required: [serviceType, rating]
    FeedbackListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Feedback"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, pagination]
...
    CitizenListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/UserProfile"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, pagination]
    AuditLog:
      type: object
      properties:
        id:
          type: string
        userId:
          type: ["string", "null"]
        action:
          type: string
        entity:
          type: string
        entityId:
          type: ["string", "null"]
        ipAddress:
          type: ["string", "null"]
        userAgent:
          type: ["string", "null"]
        metadata:
          type: ["string", "null"]
          description: JSON-encoded metadata
        createdAt:
          type: string
          format: date-time
      required: [id, action, entity, createdAt]
    AuditLogListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, pagination]
    RevenueReport:
      type: object
      properties:
        totalRevenue:
          type: number
        propertyTaxRevenue:
          type: number
        waterTaxRevenue:
          type: number
        parkingRevenue:
          type: number
        revenueByPeriod:
          type: array
          items:
            $ref: "#/components/schemas/TrendPoint"
        collectionRate:
          type: number
      required: [totalRevenue, propertyTaxRevenue, waterTaxRevenue, parkingRevenue, revenueByPeriod, collectionRate]
...